crop_pilot/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
deap==1.4.3
Django==4.2.10
djangorestframework==3.16.1
geopandas==1.1.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pyogrio==0.12.1
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
folium
